+++ pyonfx/__init__.py
@@ -2,7 +2,7 @@
 
 from .ass_core import Ass, Char, Line, Meta, Style, Syllable, Word
 from .convert import ColorModel, Convert
-from .font_utility import Font
+from .font import Font
 from .pixel import Pixel, PixelCollection
 from .shape import Shape, ShapeElement
 from .utils import ColorUtility, FrameUtility, Utils
